Empty main package and test suite
This will allow us to run `ginkgo -r` from the root of the repo while we
are making sub-packages. We'll fill in the main package later when we
create the CLI.

The empty test suite will also ensure that the main package can be
compiled, even if we don't explicitly test it's functionality.