db: Implement PutUser
Store a reference to a UAA/CF user UUID so that we can use it in a join
table between documents (terms of conditions) and agreements (people
that have agreed to terms).

We don't strictly need the table right now, because it only contains a
UUID, but we may use the table to store more information about users in
the future.

The user can be inserted many times, idempotently, so that we don't need
to do additional trips to the database to check whether it already
exists. We'd probably need to change this when we start to store more
information than the UUID.