api: Handler for GET /documents/:name
So that you can get the content of an existing document. We suspect that
this won't be used by paas-admin, because it will want to query all
agreed/unagreed documents for a user, but this will be useful for us to
debug the system without connecting directly to the database.