db: Implement basic document support
Implement putting and getting a document from the database.

We've cribbed some code from alphagov/paas-billing for setting up a
temporary database in each spec, so that there are no side effects
between specs.

We're using golang-migrate/migrate for database migrations because it
looks like a mature (well tested and well used) library. We can decide
later how/when the migrations are run in relation to the server binary.