api: Server for handlers
Attach all of the handlers that we previously created. The tests are
minimal because all of the functionality is tested directly in the
handlers.

Everything is behind basic auth, so that only ourselves and paas-admin
have access when this is deployed as an app on PaaS.

We've cribbed the code from paas-billing to handle a graceful shutdown
when CF issues a SIGTERM to stop the app.