Add username column to users table, and allow email to be nullable
Updates the API surface to match, and validation to suit.

Not all users have an email address, such as the admin user, but we want to
maintain a display name for all users. We add the username column for this
purpose.

In the future, we will need to make the username field non-nullable, but only
once other work to insert non-email-addressed users in to paas-accounts at
different times.

Allows the username to be added as part of a PATCH if it's previously set to
null.